ai_layer/requirements.txt
fastapi

uvicorn

python-dotenv

anthropic

pydantic

langchain

langchain-community

faiss-cpu

sentence-transformers

pypdf